p2p: add gray peer list housekeeping system
A random peer from the gray peer list is selected and a connection is made to check if the peer is alive.
If the connection and handshake are successful the peer is promoted to the white peer list, in case of failure the peer is evicted from the gray peer list.
The connection is closed after the check in either case.

Based on Monero solution